requirements.txt
asgiref==3.7.2
Django==4.2.9
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
environ==1.0
inflection==0.5.1
mysqlclient==2.2.1
packaging==23.2
PyJWT==2.8.0
PyMySQL==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
cryptography==41.0.7
requests==2.31.0
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
gevent==23.9.1
openai==1.7.0